VIDSOL-543 pin participant feature - detekt fixes
