+++ app/src/main/java/com/vonage/android/screen/room/components/ParticipantVideoCard.kt
@@ -3,20 +3,15 @@ package com.vonage.android.screen.room.components
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.ModalBottomSheet
@@ -30,28 +25,23 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.platform.LocalHapticFeedback
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.vonage.android.R
 import com.vonage.android.compose.components.AudioVolumeIndicator
 import com.vonage.android.compose.components.AvatarInitials
 import com.vonage.android.compose.components.ParticipantVideoRenderer
 import com.vonage.android.compose.preview.buildParticipants
 import com.vonage.android.compose.theme.VonageVideoTheme
 import com.vonage.android.compose.vivid.icons.VividIcons
-import com.vonage.android.compose.vivid.icons.solid.EndCall
 import com.vonage.android.compose.vivid.icons.solid.MicMute
 import com.vonage.android.compose.vivid.icons.solid.Microphone2
 import com.vonage.android.compose.vivid.icons.solid.MoreVertical
 import com.vonage.android.compose.vivid.icons.solid.Pin2
-import com.vonage.android.compose.vivid.icons.solid.Pin2Off
 import com.vonage.android.fx.ui.BlurIndicator
 import com.vonage.android.kotlin.model.Participant
 import com.vonage.android.kotlin.model.PublisherParticipant
@@ -158,107 +148,6 @@ fun ParticipantVideoCard(
                 isPinned = isPinned
             )
         }
-    }
-}
-
-@PreviewLightDark
-@Composable
-private fun ParticipantContextualActionsPreview() {
-    VonageVideoTheme {
-        ParticipantContextualActions(
-            participant = buildParticipants(1).first(),
-            actions = MeetingRoomActions(),
-            isPinned = true,
-        )
-    }
-}
-
-@Composable
-private fun ParticipantContextualActions(
-    participant: Participant,
-    actions: MeetingRoomActions,
-    isPinned: Boolean
-) {
-    val isMicEnabled by participant.isMicEnabled.collectAsStateWithLifecycle()
-    Column {
-        Text(
-            modifier = Modifier.padding(
-                start = VonageVideoTheme.dimens.paddingSmall,
-                bottom = VonageVideoTheme.dimens.paddingSmall,
-            ),
-            text = participant.name,
-            color = VonageVideoTheme.colors.textSecondary,
-            style = VonageVideoTheme.typography.heading2,
-        )
-
-        HorizontalDivider()
-
-        val pinnedIcon = if (isPinned) {
-            VividIcons.Solid.Pin2Off
-        } else {
-            VividIcons.Solid.Pin2
-        }
-        val pinnedLabel = if (isPinned) {
-            stringResource(R.string.meeting_room_unpin_participant)
-        } else {
-            stringResource(R.string.meeting_room_pin_participant)
-        }
-        ActionRow(
-            onClick = {
-                actions.onTogglePinParticipant(participant.id)
-            },
-            icon = pinnedIcon,
-            label = pinnedLabel,
-        )
-        if (isMicEnabled) {
-            ActionRow(
-                onClick = {
-                    actions.onForceMuteParticipant(participant.id)
-                },
-                icon = VividIcons.Solid.MicMute,
-                label = "Mute",
-            )
-        }
-        ActionRow(
-            onClick = {
-                actions.onForceMuteParticipant(participant.id)
-            },
-            icon = VividIcons.Solid.EndCall,
-            label = "Kick",
-        )
-    }
-}
-
-@Composable
-private fun ActionRow(
-    onClick: () -> Unit,
-    icon: ImageVector,
-    label: String,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .combinedClickable(
-                onClick = onClick,
-            )
-            .padding(
-                horizontal = VonageVideoTheme.dimens.paddingDefault,
-                vertical = VonageVideoTheme.dimens.paddingDefault,
-            ),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(VonageVideoTheme.dimens.spaceSmall),
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = VonageVideoTheme.colors.onSurface,
-            modifier = Modifier.size(VonageVideoTheme.dimens.iconSizeDefault),
-        )
-        Text(
-            text = label,
-            color = VonageVideoTheme.colors.textSecondary,
-            style = VonageVideoTheme.typography.bodyExtended,
-        )
     }
 }
