+++ app/src/main/java/com/vonage/android/screen/room/components/bottombar/ParticipantsList.kt
@@ -15,17 +15,15 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.vonage.android.R
 import com.vonage.android.compose.components.AvatarInitials
@@ -40,7 +38,6 @@ import com.vonage.android.compose.vivid.icons.solid.Pin2Off
 import com.vonage.android.kotlin.model.Participant
 import com.vonage.android.screen.room.MeetingRoomActions
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.ImmutableSet
 import kotlinx.collections.immutable.persistentSetOf
 import kotlinx.collections.immutable.toImmutableList
 import java.text.Normalizer
@@ -93,15 +90,16 @@ fun ParticipantsList(
                 onValueChange = { searchQuery = it },
 
                 label = { Text(text = stringResource(R.string.meeting_room_participants_list_search_placeholder)) },
-                modifier = Modifier.fillMaxWidth().testTag(SEARCH_TAG)
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .testTag(SEARCH_TAG)
             )
         }
-        if(sortedParticipants.isEmpty()){
-            item{
+        if (sortedParticipants.isEmpty()) {
+            item {
                 Text(text = stringResource(R.string.meeting_room_participants_list_search_not_found))
             }
-        }
-        else{
+        } else {
             items(
                 items = sortedParticipants,
                 key = { participant -> participant.id },
